fix: Starting to move projects over to new, more .NETy layout
Affects #157
