+++ Buttplug.Client/ButtplugClientDevice.cs
