+++ Buttplug.Client/ButtplugWSClient.cs
