+++ Buttplug.Client/LogEventArgs.cs
