+++ Buttplug.Client/Properties/AssemblyInfo.cs
