+++ Buttplug.Client/ScanningFinishedEventArgs.cs
