+++ Buttplug.Core/App_Packages/LibLog.4.2/LibLog.cs
