+++ Buttplug.Core/Buttplug.Core.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -7,8 +7,8 @@
     <ProjectGuid>{FC4AF8FE-8C21-4E07-8D24-7FB7791149D0}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ButtplugShared</RootNamespace>
-    <AssemblyName>ButtplugShared</AssemblyName>
+    <RootNamespace>Buttplug.Core</RootNamespace>
+    <AssemblyName>Buttplug.Core</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <NuGetPackageImportStamp>
@@ -31,7 +31,7 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <CodeAnalysisRuleSet>..\Buttplug\ButtplugCodeAnalysisRules.ruleset</CodeAnalysisRuleSet>
+    <CodeAnalysisRuleSet>..\Buttplug.ProjectFiles\ButtplugCodeAnalysisRules.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="JetBrains.Annotations, Version=10.4.0.0, Culture=neutral, PublicKeyToken=1010a0d8d6380325, processorArchitecture=MSIL">
@@ -53,29 +53,31 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App_Packages\LibLog.4.2\LibLog.cs" />
-    <Compile Include="Core\ButtplugConsts.cs" />
-    <Compile Include="Core\ButtplugDeviceMessage.cs" />
-    <Compile Include="Core\ButtplugJsonMessageParser.cs" />
-    <Compile Include="Core\ButtplugLog.cs" />
-    <Compile Include="Core\ButtplugLogLevel.cs" />
-    <Compile Include="Core\ButtplugLogManager.cs" />
-    <Compile Include="Core\ButtplugLogMessageEventArgs.cs" />
-    <Compile Include="Core\ButtplugMessage.cs" />
-    <Compile Include="Core\IButtplugDevice.cs" />
-    <Compile Include="Core\IButtplugLog.cs" />
-    <Compile Include="Core\IButtplugLogManager.cs" />
-    <Compile Include="Core\MessageReceivedEventArgs.cs" />
-    <Compile Include="Core\ButtplugDevice.cs" />
+    <Compile Include="ButtplugConsts.cs" />
+    <Compile Include="ButtplugDeviceMessage.cs" />
+    <Compile Include="ButtplugJsonMessageParser.cs" />
+    <Compile Include="ButtplugLog.cs" />
+    <Compile Include="ButtplugLogLevel.cs" />
+    <Compile Include="ButtplugLogManager.cs" />
+    <Compile Include="ButtplugLogMessageEventArgs.cs" />
+    <Compile Include="ButtplugMessage.cs" />
+    <Compile Include="IButtplugDevice.cs" />
+    <Compile Include="IButtplugLog.cs" />
+    <Compile Include="IButtplugLogManager.cs" />
+    <Compile Include="MessageReceivedEventArgs.cs" />
+    <Compile Include="ButtplugDevice.cs" />
     <Compile Include="Messages\Messages.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="..\dependencies\buttplug-schema\schema\buttplug-schema.json" />
-    <AdditionalFiles Include="..\stylecop.json" />
     <None Include="app.config" />
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
+  </ItemGroup>
+  <ItemGroup>
+    <AdditionalFiles Include="..\Buttplug.ProjectFiles\stylecop.json" />
   </ItemGroup>
   <ItemGroup>
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
