+++ Buttplug.Core/ButtplugConsts.cs
