+++ Buttplug.Core/ButtplugDeviceMessage.cs
