+++ Buttplug.Core/ButtplugLogLevel.cs
