+++ Buttplug.Core/ButtplugLogManager.cs
