+++ Buttplug.Core/ButtplugLogMessageEventArgs.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Buttplug.Messages;
+using Buttplug.Core.Messages;
 using JetBrains.Annotations;
 
 namespace Buttplug.Core
