+++ Buttplug.Core/ButtplugMessage.cs
