+++ Buttplug.Core/IButtplugDevice.cs
