+++ Buttplug.Core/IButtplugLog.cs
@@ -1,6 +1,6 @@
-﻿using Buttplug.Messages;
+﻿using Buttplug.Core.Messages;
 using JetBrains.Annotations;
-using static Buttplug.Messages.Error;
+using static Buttplug.Core.Messages.Error;
 
 namespace Buttplug.Core
 {
