+++ Buttplug.Core/IButtplugLogManager.cs
