+++ Buttplug.Core/Properties/AssemblyInfo.cs
