+++ Buttplug.ProjectFiles/ButtplugCodeAnalysisRules.ruleset
