+++ Buttplug.Server/Bluetooth/BluetoothSubtypeManager.cs
