+++ Buttplug.Server/Bluetooth/ButtplugBluetoothDevice.cs
