+++ Buttplug.Server/Bluetooth/Devices/FleshlightLaunch.cs
