+++ Buttplug.Server/Bluetooth/Devices/Kiiroo.cs
