+++ Buttplug.Server/Bluetooth/Devices/Lovense.cs
