+++ Buttplug.Server/Bluetooth/IBluetoothDeviceInfo.cs
