+++ Buttplug.Server/Bluetooth/IBluetoothDeviceInterface.cs
