+++ Buttplug.Server/DeviceAddedEventArgs.cs
