+++ Buttplug.Server/DeviceManager.cs
