+++ Buttplug.Server/DeviceSubtypeManager.cs
