+++ Buttplug.Server/IButtplugServiceFactory.cs
