+++ Buttplug.Server/IDeviceSubtypeManager.cs
