+++ Buttplug.Server/Properties/AssemblyInfo.cs
